# src/ActionLedger.Web/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ActionLedger API",
    "description": "Turns meeting notes into tracked, attributable action items. Every route lives under /api/v1 except /health, /openapi, and /swagger. Errors are RFC 9457 ProblemDetails whose type is one of validation, unauthorized, forbidden, not-found, or conflict.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/ai/provider": {
      "get": {
        "tags": [
          "Ai"
        ],
        "summary": "Reports the active AI provider and model.",
        "description": "provider is the configured Ai:Provider (Fake, LocalOpenAI or AzureOpenAI) and model is what that provider reports. No endpoint, key or base URL is published.",
        "operationId": "GetAiProvider",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProviderDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiProviderDto": {
        "required": [
          "provider",
          "model"
        ],
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "An HS256 token from POST /api/v1/auth/login carrying sub, name, and role. Send it as: Authorization: Bearer {token}.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}